Moved resize_images utility fn to mask module
