+++ mask/eolearn/mask/__init__.py
@@ -4,6 +4,7 @@ Public classes and functions of mask subpackage
 
 from .cloud_mask import AddCloudMaskTask, get_s2_pixel_cloud_detector
 from .masking import AddValidDataMaskTask, MaskFeature
+from .utilities import resize_images
 
 
 __version__ = '0.5.0'
